wip: use aws example hooks
