+++ apps/af-mvp/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@aws-sdk/client-ssm": "^3.352.0",
     "af-shared": "*",
+    "amazon-cognito-passwordless-auth": "^0.12.1",
     "aws-jwt-verify": "^4.0.0",
     "dotenv": "^16.3.1",
     "jsonwebtoken": "^9.0.0",
