+++ apps/af-mvp/src/pages/auth/sign-in/components/sign-in.tsx
@@ -1,130 +1,51 @@
-import { useState } from 'react';
-import { SubmitHandler, useForm } from 'react-hook-form';
+import { useRouter } from 'next/router';
+import { useEffect, useState } from 'react';
+import { usePasswordless } from 'amazon-cognito-passwordless-auth/react';
 import { Button, Text } from 'suomifi-ui-components';
-import FormInput from '@shared/components/form/form-input';
+import Page from '@shared/components/layout/page';
 import CustomHeading from '@shared/components/ui/custom-heading';
 import Loading from '@shared/components/ui/loading';
+import SignInForm from './sign-in-form';
 
-interface SubmitProps {
-  text: string;
-  isSubmitting: boolean;
-}
+export default function SingInPage() {
+  const { signInStatus } = usePasswordless();
+  const router = useRouter();
+  const [isLoading, setIsLoading] = useState(true);
 
-interface FormProps {
-  handleFormSubmit: (value: string) => void;
-}
+  if (signInStatus == 'SIGNED_IN') {
+    return (
+      <div className="flex flex-col items-center justify-center mt-8 gap-6">
+        Kill your cognito session
+        <Button onClick={() => alert('no can do')}>Kill it</Button>
+      </div>
+    );
+  }
 
-type EmailForm = { email: string };
-type CodeForm = { code: string };
-
-function Submit({ text, isSubmitting }: SubmitProps) {
   return (
-    <div className="flex flex-row gap-3 items-center">
-      <Button type="submit">{text}</Button>
-      {isSubmitting && (
-        <div className="mt-1 ml-1">
-          <Loading variant="small" />
+    <Page title="Sign In">
+      <Page.Block className="bg-white">
+        <div className="flex flex-col mt-8 gap-6">
+          <div className="bg-suomifi-blue-bg-light p-4 flex flex-col gap-6">
+            <CustomHeading variant="h3" className="!text-lg">
+              Sing in with email
+            </CustomHeading>
+            <Text>
+              We utilize a one-time code verification process. Please enter your
+              email address below to receive a unique code that you can use to
+              sign in securely.
+            </Text>
+            <Text>
+              Once you’ve entered your email address, a one-time code will be
+              sent to your inbox. Use that code to complete the sign-in process.
+              Your privacy is important to us. We will only use your email
+              address for the purpose of sending the verification code and
+              managing your account security.
+            </Text>
+          </div>
+
+          <SignInForm />
         </div>
-      )}
-    </div>
-  );
-}
-
-function EmailForm({ handleFormSubmit }: FormProps) {
-  const {
-    handleSubmit,
-    control,
-    formState: { isSubmitting },
-  } = useForm<EmailForm>();
-
-  const onSubmit: SubmitHandler<EmailForm> = async ({ email }) => {
-    await handleFormSubmit(email);
-  };
-
-  return (
-    <form onSubmit={handleSubmit(onSubmit)} className="flex flex-col gap-6">
-      <div className="flex flex-col gap-4 items-start">
-        <CustomHeading variant="h4">Log in</CustomHeading>
-        <Text>
-          Enter your email address you use in Access Finland. We will send you a
-          verification code to your email address.
-        </Text>
-        <FormInput
-          name="email"
-          type="email"
-          labelText="Your email address"
-          control={control}
-          rules={{ required: 'Email is required' }}
-        />
-        <Submit text="Send me a code" isSubmitting={isSubmitting} />
-      </div>
-    </form>
-  );
-}
-
-function CodeForm({ handleFormSubmit }: FormProps) {
-  const {
-    handleSubmit,
-    control,
-    formState: { isSubmitting },
-  } = useForm<CodeForm>();
-
-  const onSubmit: SubmitHandler<CodeForm> = async ({ code }) => {
-    await handleFormSubmit(code);
-  };
-
-  return (
-    <form onSubmit={handleSubmit(onSubmit)} className="flex flex-col gap-6">
-      <div className="flex flex-col gap-4 items-start">
-        <CustomHeading variant="h4">Enter the vefication code</CustomHeading>
-        <Text>
-          If you did not receive a verification code in your email, please check
-          your spam folder. Also make sure you entered the same email address
-          you used to sign in to Access Finland.
-        </Text>
-        <FormInput
-          name="code"
-          type="text"
-          labelText="Your code"
-          // eslint-disable-next-line jsx-a11y/no-autofocus
-          autoFocus
-          control={control}
-          rules={{ required: 'Code is required' }}
-        />
-        <Submit text="Sign in with code" isSubmitting={isSubmitting} />
-      </div>
-    </form>
-  );
-}
-
-const sleep = () => new Promise(resolve => setTimeout(resolve, 1500));
-
-export default function SignIn() {
-  const [isCodeSent, setCodeSent] = useState(false);
-
-  const handleEmailSubmit = async (email: string) => {
-    await sleep();
-    setCodeSent(true);
-  };
-
-  const handleCodeSubmit = async (code: string) => {
-    await sleep();
-    console.log(code);
-  };
-
-  return (
-    <div className="flex flex-col gap-4 items-start">
-      {!isCodeSent ? (
-        <EmailForm handleFormSubmit={handleEmailSubmit} />
-      ) : (
-        <CodeForm handleFormSubmit={handleCodeSubmit} />
-      )}
-      <button
-        className="text-blue-600 hover:text-blue-800 visited:text-purple-600 !text-base"
-        onClick={() => setCodeSent(!isCodeSent)}
-      >
-        {isCodeSent ? 'I need a new code' : 'I already have a code'}
-      </button>
-    </div>
+      </Page.Block>
+    </Page>
   );
 }
